+++ tests/Fixtures/Post.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace JacobJoergensen\LaravelPaper\Tests\Fixtures;
 
+use Illuminate\Database\Eloquent\Attributes\Scope;
 use Illuminate\Database\Eloquent\Model;
 use JacobJoergensen\LaravelPaper\Attributes\ContentPath;
 use JacobJoergensen\LaravelPaper\Attributes\Driver;
@@ -19,6 +20,12 @@ final class Post extends Model
     public function scopePublished(PaperQueryBuilder $query): PaperQueryBuilder
     {
         return $query->where('published', true);
+    }
+
+    #[Scope]
+    protected function withOrder(PaperQueryBuilder $query, int $order): PaperQueryBuilder
+    {
+        return $query->where('order', $order);
     }
 
     public function author(): ?Author